requirements.txt
tk
reportlab
pillow
tkcalendar
pg8000
python-dotenv
pyperclip
pywin32
